# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: SGG Digital API
  description: |
    API du Secrétariat Général du Gouvernement du Gabon — Plateforme digitale.
    
    ## Modules
    
    | Module | Description |
    |--------|-------------|
    | **Auth** | Authentification JWT (login, register, profil) |
    | **GAR** | Gestion Axée sur les Résultats — PAG 2026 |
    | **Reporting** | Matrice de reporting mensuel des ministères |
    | **Nominations** | Gestion des nominations ministérielles |
    | **Législatif** | Cycle législatif (textes de loi, promulgation) |
    | **e-GOP** | Conseil des Ministres (ordres du jour, PV) |
    | **JO** | Journal Officiel de la République |
    | **PTM** | Programme de Transformation du Monde rural |
    | **Institutions** | Référentiel des institutions gouvernementales |
    | **Monitoring** | Réception des événements de monitoring frontend |
    
    ## Authentification
    
    Toutes les routes protégées nécessitent un JWT Bearer token :
    ```
    Authorization: Bearer <token>
    ```
    
    ## Environnements
    
    | Environnement | URL |
    |---------------|-----|
    | Production | `https://api.sgg.ga` |
    | Staging | `https://staging-api.sgg.ga` |
    | Développement | `http://localhost:8080` |
    
  version: "2.1.0"
  contact:
    name: OKA Tech
    email: dev@okatech.ga
  license:
    name: Propriétaire
    identifier: UNLICENSED
servers:
  - url: https://api.sgg.ga
    description: Production
  - url: https://staging-api.sgg.ga
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenu via POST /api/auth/login
  schemas:
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "NOT_FOUND"
            message:
              type: string
              example: "Ressource non trouvée"
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "admin@sgg.ga"
        password:
          type: string
          format: password
          example: "SecureP@ss123"
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            token:
              type: string
              description: JWT token
            user:
              $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
          example: "Jean-Pierre NZOGHE"
        role:
          $ref: '#/components/schemas/AppRole'
        institution_id:
          type: string
          format: uuid
        is_active:
          type: boolean
    AppRole:
      type: string
      enum:
        - admin_sgg
        - directeur_sgg
        - sg_ministere
        - sgpr
        - premier_ministre
        - ministre
        - assemblee
        - senat
        - conseil_etat
        - cour_constitutionnelle
        - dgjo
        - citoyen

    # ── GAR ───────────────────────────────────────────────────────────────
...
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Connexion utilisateur
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Connexion réussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Identifiants incorrects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Inscription d'un nouvel utilisateur
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, full_name, role]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                full_name:
                  type: string
                role:
                  $ref: '#/components/schemas/AppRole'
      responses:
        '201':
          description: Compte créé
        '409':
          description: Email déjà existant
...